Remove uiframework compile dependency on the (non open source) datagrabber module. The dependency was some constants from DGPSPConstants (referencing constants from ConveyancePrototype) to describe Asset Types (it is a Ship? Plane? Train?).
Solve the dependency by creating a new constants file that copies the
constants. Comments in the datagrabber module warn future developers
to please update the constants here if they change in the datagrabber
module.

Note that this removes the compile dependency, but actual usage
probably still requires the datagrabber module.